Fixes issue 60 http://code.google.com/p/winforms-geplugin-control-library/issues/detail?id=60
Fixes issue 63 http://code.google.com/p/winforms-geplugin-control-library/issues/detail?id=63

Rewrote the KmlTreeView and KmlTreeViewNode classes to make the display virtual (so that node conent is only loaded on demand)

Replaced the ApiType constants with an Enum of the same.

Removed the StringEnums classes and added a Helper/Extensions class to handle the string descriptions in the Enumerations/Layer.cs

Changed the event handling in the GEWebBrowser class to use delegates.

Moved from v2 to v3 of the Google Maps Api for Geocoding 

Plus a whole host of other tweaks and fixes (mainly GEServer and HttpRequest) - phew!